requirements.txt
discord.py
DiscordUtils
asyncpg
flask
discord-flags
jishaku
praw
humanize
